heroku_app/intellomedix final version/requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.0.2
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.15.4
olefile==0.46
Pillow==5.3.0
PyYAML==5.3.1
six==1.15.0
https://download.pytorch.org/whl/nightly/cpu/torch-1.3.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchvision==0.4.2
Werkzeug==1.0.1
wincertstore==0.2
gunicorn==19.9.0
psycopg2==2.8.5
SQLAlchemy==1.3.18
Django==2.2.5
Flask-SQLAlchemy==2.4.4
pytz==2019.2
scipy==1.4.1
sqlparse==0.3.0
opencv-python==4.3.0.36
psycopg2==2.8.5
pydicom==2.0.0
pypng==0.0.20
SQLAlchemy==1.3.18
Werkzeug==1.0.1











